feat: Set up basic infrastructure
This commit sets up the basic infrastructure for the project. It includes a new blank `.env-none` file, a `.env.template`, a `.gitignore` file that ignores a new `no_commit` folder, a `.husky/pre-commit` file that restores an original comment line, and a few additional setup files such as `.tool-versions` and `.vscode` configurations.

Also, this commit includes a change to the parent repository's `full-test.yml` workflow file so that the `test` job executes only when a branch is created, reopened, or synchronized, and on the develop and staging branches, but not on the main branch.